Allow slash-infra to search multiple AWS accounts!
It currently does the searching sequentially, but this is a good POC.

This approach will require reworking if we want to APIs outside of EC2.
When that happens we'll probably want to re-use instances of the STS
credentials for each account's API client.

Also taking a simplistic approach to supporting multiple regions - 90%
of our stuff is in one region, so it makes sense for other regions to be
an exception, and search only one AWS region by default.